+++ spring-data-eclipse-store/src/test/java/software/xdev/spring/data/eclipse/store/helper/DummyEntityProvider.java
@@ -19,7 +19,8 @@ import java.util.Collection;
 import java.util.List;
 
 import software.xdev.spring.data.eclipse.store.core.EntityProvider;
-import software.xdev.spring.data.eclipse.store.repository.root.EntityData;
+import software.xdev.spring.data.eclipse.store.repository.root.v2_4.EntityData;
+import software.xdev.spring.data.eclipse.store.repository.root.v2_4.NonLazyEntityData;
 
 
 public class DummyEntityProvider<T> extends EntityProvider<T, Void>
@@ -27,9 +28,9 @@ public class DummyEntityProvider<T> extends EntityProvider<T, Void>
 	public DummyEntityProvider(final Collection<T> collection)
 	{
 		super();
-		final EntityData<T, Void> objects = new EntityData<>();
-		objects.setIdGetter(i -> null);
-		objects.getEntities().addAll(collection);
+		final EntityData<T, Void> objects = new NonLazyEntityData<>();
+		objects.setIdGetter(null);
+		collection.forEach(objects::ensureEntityAndReturnObjectsToStore);
 		this.addEntityData(objects);
 	}