+++ spring-data-eclipse-store/src/main/java/software/xdev/spring/data/eclipse/store/core/EntityListProvider.java
@@ -18,6 +18,4 @@ package software.xdev.spring.data.eclipse.store.core;
 public interface EntityListProvider
 {
 	<T, ID> EntityProvider<T, ID> getEntityProvider(final Class<T> clazz);
-	
-	<T> long getEntityCount(final Class<T> clazz);
 }