+++ spring-data-eclipse-store/src/main/java/software/xdev/spring/data/eclipse/store/repository/root/VersionedRoot.java
@@ -19,15 +19,20 @@ import software.xdev.micromigration.version.MigrationVersion;
 import software.xdev.micromigration.version.Versioned;
 import software.xdev.spring.data.eclipse.store.repository.EclipseStoreMigrator;
 import software.xdev.spring.data.eclipse.store.repository.Root;
+import software.xdev.spring.data.eclipse.store.repository.root.v2_4.RootDataV2_4;
 
 
+@SuppressWarnings("deprecation")
 public class VersionedRoot implements Versioned
 {
 	private MigrationVersion version;
 	
 	private Root rootDataV1;
 	
-	private final RootDataV2 rootDataV2;
+	private RootDataV2 rootDataV2;
+	
+	@SuppressWarnings("checkstyle:MemberName")
+	private final RootDataV2_4 rootDataV2_4;
 	
 	public VersionedRoot()
 	{
@@ -40,7 +45,7 @@ public class VersionedRoot implements Versioned
 	public VersionedRoot(final Root rootDataV1)
 	{
 		this.rootDataV1 = rootDataV1;
-		this.rootDataV2 = new RootDataV2();
+		this.rootDataV2_4 = new RootDataV2_4();
 		if(rootDataV1 != null)
 		{
 			this.version = new MigrationVersion(0, 0, 0);
@@ -56,19 +61,36 @@ public class VersionedRoot implements Versioned
 		return this.rootDataV1;
 	}
 	
+	/**
+	 * @deprecated and is only used in tests
+	 */
+	@SuppressWarnings("DeprecatedIsStillUsed")
+	@Deprecated
+	public void setRootDataV2(final RootDataV2 rootDataV2)
+	{
+		this.rootDataV2 = rootDataV2;
+	}
+	
 	public RootDataV2 getRootDataV2()
 	{
 		return this.rootDataV2;
 	}
 	
+	@SuppressWarnings("checkstyle:MethodName")
+	public RootDataV2_4 getRootDataV2_4()
+	{
+		return this.rootDataV2_4;
+	}
+	
 	public void clearOldRootData()
 	{
 		this.rootDataV1 = null;
+		this.rootDataV2 = null;
 	}
 	
-	public RootDataV2 getCurrentRootData()
+	public RootDataV2_4 getCurrentRootData()
 	{
-		return this.rootDataV2;
+		return this.rootDataV2_4;
 	}
 	
 	@Override