+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Updated org.springframework.boot.version to v3.4.0
 * Updated EclipseStore to v2.0.0
+* Implemented Lazy Repositories with ``LazyEclipseStoreRepository``
 
 # 2.3.1